update comments and current comment

+++ app/src/main/java/com/bunbeauty/ideal/myapplication/cleanArchitecture/business/commets/creationComment/CreationCommentMessageInteractor.kt
@@ -18,7 +18,7 @@ class CreationCommentMessageInteractor(
 
 
     override fun checkMessage(
-        rating: Double,
+        rating: Float,
         review: String,
         creationCommentPresenterCallback: CreationCommentPresenterCallback
     ) {
