+++ app/src/main/java/com/bunbeauty/ideal/myapplication/cleanArchitecture/business/commets/creationComment/iCreationComment/ICreationCommentMessageInteractor.kt
@@ -16,7 +16,7 @@ interface ICreationCommentMessageInteractor {
     )
 
     fun checkMessage(
-        rating: Double,
+        rating: Float,
         review: String,
         creationCommentPresenterCallback: CreationCommentPresenterCallback
     )
