+++ app/src/main/java/com/bunbeauty/ideal/myapplication/cleanArchitecture/business/commets/creationComment/iCreationComment/ICreationCommentServiceCommentInteractor.kt
@@ -4,7 +4,7 @@ import com.bunbeauty.ideal.myapplication.cleanArchitecture.callback.comments.Cre
 import com.bunbeauty.ideal.myapplication.cleanArchitecture.data.db.models.entity.comment.ServiceComment
 
 interface ICreationCommentServiceCommentInteractor {
-    fun setRating(rating: Double)
+    fun setRating(rating: Float)
     fun setReview(review: String)
     fun createServiceComment(
         serviceComment: ServiceComment,
